MODWRKFLOW-34: Update the group ID and topic pattern to be more consistent.
The patterns should use `workflow-engine` for the engine.
The engine happens to be Camunda in this particular case but this path should not be tied to Camunda.

The module name should be in there as well, which would be `mod-workflow`.
The combination does not need to be `mod-workflow-workflow-engine` as that seems silly.
Instead, the second `-workflow` is dropped.