MODWRKFLOW-34: Revert path changes and related changes.
The Camunda base path and the Camunda reset path are needed.
When going through OKAPI REST end points, these paths represent the OKAPI REST path.
This is not necessarily a path to where mod-camunda actually listens on.

This commit puts the code back in a state where the Camunda OKAPI REST paths are less hardcoded.

Remove the code smells regarding logging an error and throwing an exception at the same time that was introduced when the path changes were made.

Update the debug log to show both the tenant and token regardless of whether or not they are NULL.
Also use a single message to reduce the amount of log spamming that can happen due to this being called on every request.

Update the `application.yml` from the tests to more closely match the `application.yml` from the normal build.

The changes being reverted introduced problems and are outside the scope of a change from ActiveMQ to Kafka.
The `restPath` variables should probably be refactored to the `workflow-engine` rather than `rest`.
However, this has been out of scope and there are other issues set aside for handling this.