MODWRKFLOW-43: Switch to static list of classes, and rename DeserializeAsJsonResolver.
Switch to a hard-coded list of classes for conversion to reduce security concerns of a more maintainable and dynamic approach.
Use the classes in a list to help catch automatic refactoring should there be changes in the future to class names.
Adding the `extends Node` helps make this list more maintainable by preventing all non-Node classes from appearing in the list.

The `DeserializeAsJsonResolver` is specifically intended to be used for `deserializeAs` on a `Node`.
Change the class name to better communicate this.
Add java documentation comments to the class.

Update the unit tests accordingly.
There are also classes within the model directory that do not extend Node.
Add some unit tests to identify a couple of these and ensure that they do not get converted.